# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin 和 GORM 构建的高性能社交网络平台接口文档",
        "title": "社交网络平台API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API 支持团队",
            "url": "http://www.swagger.io/support",
            "email": "caribdis@163.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "按时间倒序获取所有好友的动态(每次10条)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "动态"
                ],
                "summary": "获取好友动态列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "偏移量(默认0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取动态列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ParamPostWithUserInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建用户动态(支持纯文本、纯图片、图文混合)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "动态"
                ],
                "summary": "创建用户动态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文字内容(不超过500字)",
                        "name": "content",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "图片文件(最多9张,支持jpg/jpeg/png)",
                        "name": "images",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "动态创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Post"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误/图片格式错误/图片过多",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "413": {
                        "description": "文件过大",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/posts/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除用户自己的动态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "动态"
                ],
                "summary": "删除动态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "动态ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/posts/{id}/view": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "每次浏览动态时增加1次浏览量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "动态"
                ],
                "summary": "增加动态浏览量",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "动态ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/posts/{user_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "按时间倒序获取指定用户的动态(每次10条)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "动态"
                ],
                "summary": "获取指定用户动态列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "偏移量(默认0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取用户动态",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ParamPostWithUserInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "非好友关系",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ParamPostWithUserInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "用户头像",
                    "type": "string"
                },
                "content": {
                    "description": "文字内容",
                    "type": "string"
                },
                "created_at": {
                    "description": "发布时间",
                    "type": "string"
                },
                "images": {
                    "description": "图片URL，多个用逗号分隔",
                    "type": "string"
                },
                "nickname": {
                    "description": "用户备注或用户名",
                    "type": "string"
                },
                "view_count": {
                    "description": "浏览量",
                    "type": "integer"
                }
            }
        },
...
        "models.Post": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "description": "发布用户头像",
                    "type": "string"
                },
                "content": {
                    "description": "文字内容",
                    "type": "string"
                },
                "created_at": {
                    "description": "发布时间",
                    "type": "string"
                },
                "images": {
                    "description": "图片URL，多个用逗号分隔",
                    "type": "string"
                },
                "user_id": {
                    "description": "发布用户ID",
                    "type": "string",
                    "example": "0"
                },
                "username": {
                    "description": "发布用户昵称",
                    "type": "string"
                },
                "view_count": {
                    "description": "浏览量",
                    "type": "integer"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "业务数据（如UID字符串）"
                },
                "msg": {
                    "description": "消息描述",
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}